+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## PENDING - 0.1.0
 
+Changes:
+
+- The HTTP health check method now uses [HTTPX](https://www.python-httpx.org/) instead of `requests` to make the request. With this change HTTP2 support is now available for the health check (defaults to False).
+- HTTP check SNI adapter removed - HTTPX can handle SNI natively without requiring an adapter
+
 Misc:
 
 - Update Sentry, Apprise and Pydantic releases